==> scripts/06_diffAnalysis/23_plotREVIGO.R <==
######################################################################
### Goal: Edit REVIGO plots 
### Author: Janay Fox
### R script
#####################################################################

## Set up ##
library(ggplot2)
library(scales)
library(dplyr)
library(viridis)
library(ggrepel)
library(patchwork)

#load in data
BN.mfuzz.cluster6.revigo.names <- c("term_ID","description","frequency","plot_X","plot_Y","log_size","value","uniqueness","dispensability");
BN.mfuzz.cluster6.revigo.data <- rbind(c("GO:0006575","cellular modified amino acid metabolic process",0.6851865747895725,-1.4626573931674696,-7.375957401785821,5.313865112138396,-3.275537877533078,0.8045735065486667,0.12189312),
                     c("GO:0018126","protein hydroxylation",0.014359122143742097,-4.660596939435513,0.46732149939654,3.635282637998212,-5.976993696766804,0.7625601204623326,0.07550465),
                     c("GO:0018208","peptidyl-proline modification",0.20196901915454343,-4.437312255783912,-2.4199390646681382,4.783346067476675,-5.06095941946776,0.6636365526335498,0.48991279),
                     c("GO:0019471","4-hydroxyproline metabolic process",0.00013969959000166042,5.223544575637416,-2.965993957712103,1.6334684555795864,-6.709255333356367,0.8080486425796953,0),
                     c("GO:0019511","peptidyl-proline hydroxylation",0.008737876736532427,-5.962441729143965,-1.722079985237315,3.419625360887743,-6.5960949160448035,0.623059453497824,0.24916865),
                     c("GO:0071456","cellular response to hypoxia",0.013477684254445906,1.9735185641624111,5.149955091525602,3.607776603741693,-2.1106418862833567,0.6559244350235264,0.00436528),
                     c("GO:1901605","alpha-amino acid metabolic process",3.5585210895822956,1.495944422195505,-3.9409029151515487,6.029324108830442,-2.904129865219576,0.66522545048598,0.32679993));

BN.mfuzz.cluster5.revigo.names <- c("term_ID","description","frequency","plot_X","plot_Y","log_size","value","uniqueness","dispensability");
BN.mfuzz.cluster5.revigo.data <- rbind(c("GO:0002252","immune effector process",0.060752690747150666,-5.998094700820335,-4.598450761702338,4.26164345331189,-300,0.6216655149410969,0.68371319),
                     c("GO:0002376","immune system process",0.5358776487042264,-0.2775107065212202,6.449903227701661,5.207122497650964,-4.375068085808162,1,-0),
                     c("GO:0002682","regulation of immune system process",0.29494907484279137,0.38160185579835487,-6.426453588376232,4.947806094714132,-2.3318140635124465,0.8503288591443282,0.12346543),
                     c("GO:0006952","defense response",1.0410213876080876,-7.070536298917201,1.6169622180610175,5.495515198600499,-3.0263368029776636,0.6752063450058148,0.26469029),
                     c("GO:0006956","complement activation",0.02405161274528587,-5.15145922938502,-2.4184092854987003,3.859258417467307,-300,0.24511752745717946,0.68120383),
                     c("GO:0006959","humoral immune response",0.040652580690483185,-6.3910942510337945,-1.7782713508850148,4.087177811761157,-300,0.5212241731017286,-0),
                     c("GO:0009617","response to bacterium",0.14606922606959327,-5.82063423423205,2.629817287853001,4.642622776409274,-2.8459599313518695,0.5644667023925283,0.29520673),
                     c("GO:0072376","protein activation cascade",0.0026675969328888494,3.8045675074541405,2.436255274690121,2.904715545278681,-300,1,0),
                     c("GO:2000257","regulation of protein activation cascade",0.00020954938500249067,3.6730697752147288,-3.2658876857394574,1.806179973983887,-2.5539372080910536,0.896905571765112,-0));


BA.mfuzz.cluster5.revigo.names <- c("term_ID","description","frequency","plot_X","plot_Y","log_size","value","uniqueness","dispensability");
BA.mfuzz.cluster5.revigo.data <- rbind(c("GO:0001666","response to hypoxia",0.03421642100826383,-4.0564464713848185,5.205032264723029,4.012330955580147,-1.9071272350098294,0.7168140071821428,0.13752985),
                     c("GO:0002412","antigen transcytosis by M cells in mucosal-associated lymphoid tissue",1.3304722857300992E-05,-6.0776870834810905,1.269066500353262,0.6989700043360189,-2.4567900040092248,0.8291741136955408,-0),
                     c("GO:0018126","protein hydroxylation",0.014359122143742097,0.1285268881145624,-6.58437813043557,3.635282637998212,-1.698099181216368,0.9320867795910414,0.24916865),
                     c("GO:0018208","peptidyl-proline modification",0.20196901915454343,1.0901376541411922,-6.284790046967555,4.783346067476675,-1.3344787881880034,0.8955853097897764,0.48991279),
                     c("GO:0019471","4-hydroxyproline metabolic process",0.00013969959000166042,-3.9188251835338996,-7.11707987870439,1.6334684555795864,-1.9272992614619782,0.9749626429166117,-0),
                     c("GO:0019511","peptidyl-proline hydroxylation",0.008737876736532427,1.6548523068076866,-6.256654004212935,3.419625360887743,-1.9071272350098294,0.867255815242152,0.07386162),
                     c("GO:0032364","intracellular oxygen homeostasis",0.0007716739257234576,0.7042894830821558,-0.7491130199456139,2.367355921026019,-2.719521279912308,1,-0),
                     c("GO:0045056","transcytosis",0.0023283265000276737,-7.038519184452494,-0.0193563275157307,2.8457180179666586,-1.834765941803245,0.8673164785002478,0.35633827),
                     c("GO:0051342","regulation of cyclic-nucleotide phosphodiesterase activity",0.0004756438421485105,5.56266520482584,-2.342156089961538,2.1583624920952498,-2.321121859120142,0.938259661606441,0.42930321),
                     c("GO:0051344","negative regulation of cyclic-nucleotide phosphodiesterase activity",0.0001596566742876119,5.404183837073713,-2.941804983460242,1.6901960800285136,-2.719521279912308,0.9393948247648263,0),
                     c("GO:0055008","cardiac muscle tissue morphogenesis",0.008621460411531045,-6.2569946270004895,-1.8258014961697053,3.4138025167693513,-1.906509695903187,0.6803546744097171,0.63730513),
                     c("GO:0060343","trabecula formation",0.002624356583602621,-6.146905667207981,-3.5975843796489153,2.8976270912904414,-1.906509695903187,0.7532080836154397,0.49804856),
                     c("GO:0060711","labyrinthine layer development",0.00529860587792012,-6.611596192845193,-1.5717684519346242,3.2024883170600935,-2.2162621937396665,0.7604810534603007,0.28466572),
                     c("GO:0070482","response to oxygen levels",0.038207837865454126,-4.4397523317077185,5.289418276049121,4.060244426898225,-1.7862744074685424,0.7848039330322394,0.65312022),
                     c("GO:0071731","response to nitric oxide",0.0014236053457312061,-2.2464802711517016,6.132690705384817,2.6324572921847245,-1.906509695903187,0.900363156703191,0.17110639),
                     c("GO:0099159","regulation of modification of postsynaptic structure",0.0007151288535799284,2.24252873665915,6.508583064952034,2.3344537511509307,-2.628745250794901,0.9625634410252221,0.08962698),
                     c("GO:0140252","regulation protein catabolic process at postsynapse",0.0003791846014330783,1.3076389718485109,3.6345422073414815,2.060697840353612,-2.586088988152483,0.9644350427350025,0.09591616),
                     c("GO:1905289","regulation of CAMKK-AMPK signaling cascade",0.00038251078214740356,5.181074105828184,2.783842343395377,2.0644579892269186,-1.7981922431626012,0.9206081785744782,0.10402292),
                     c("GO:1905290","negative regulation of CAMKK-AMPK signaling cascade",6.652361428650496E-06,5.369031897330604,2.3319599157407804,0.47712125471966244,-1.906509695903187,0.924652616247932,0.63661272));

BA.snp.revigo.names <- c("term_ID","description","frequency","plot_X","plot_Y","log_size","value","uniqueness","dispensability");
BA.snp.revigo.data <- rbind(c("GO:0000002","mitochondrial genome maintenance",0.06575433816451984,-3.2192511269524613,0.028052743702160834,4.426169547035326,-2.275724130399211,0.9625742724577147,0.25223306),
                     c("GO:0000082","G1/S transition of mitotic cell cycle",0.054207030484004964,3.7371240817266567,-8.186798257958266,4.342304220718211,-1.9090369234042683,0.9953646268501949,0.00807669),
                     c("GO:0000472","endonucleolytic cleavage to generate mature 5'-end of SSU-rRNA from (SSU-rRNA, 5.8S rRNA, LSU-rRNA)",0.02006052021594673,-4.197586423472197,0.8071935408562257,3.910624404889201,-1.4594202834955456,0.9359008105152365,0.2428244),
                     c("GO:0001798","positive regulation of type IIa hypersensitivity",0.00015281389032911874,7.489077293547283,-1.1700688398897516,1.7993405494535817,-1.7977842241988684,0.7707091421889217,0.63196545),
                     c("GO:0001839","neural plate morphogenesis",0.0008256879558105609,-1.0785721995423876,-6.502004725786722,2.526339277389844,-1.4222784754909792,0.9027576437643561,0.63874141),
                     c("GO:0001934","positive regulation of protein phosphorylation",0.14391371358753086,5.977359038445358,-2.345761620124707,4.766338475251287,-2.489454989793388,0.730811092782544,0.65289523),
                     c("GO:0001947","heart looping",0.011379705349186147,-0.817050462333182,-6.7849483778209745,3.6644539285811577,-2.073143291050308,0.8740227063817231,0.56205142),
                     c("GO:0001958","endochondral ossification",0.003088319428748157,-0.7240367408253657,-6.9470913970989825,3.0982975364946976,-1.4766437933452072,0.9061987388032373,0.58024668),
                     c("GO:0002191","cap-dependent translational initiation",0.00494426877419697,-6.024079157254438,0.3881493796266677,3.3025473724874854,-1.4594202834955456,0.9610357183409,0.22191619),
                     c("GO:0002841","negative regulation of T cell mediated immune response to tumor cell",0.00021443239449408597,7.1856705747902225,2.6827292807516248,1.9444826721501687,-1.5956795327782693,0.8038559252174228,0.53722851),
                     c("GO:0003138","primary heart field specification",0.0002095029141608886,-0.5821617303348348,-6.797495862557851,1.9344984512435677,-1.6528652170899802,0.9081770339839528,0.64474909),
                     c("GO:0003185","sinoatrial valve morphogenesis",1.4788440999592136E-05,-0.21803748161534472,-6.880609471129721,0.8450980400142568,-1.7191940716063332,0.9214705936097071,0.43001535),
                     c("GO:0003365","establishment of cell polarity involved in ameboidal cell migration",0.005084758963693096,-4.62375794441537,-5.062038503669373,3.314709692955174,-1.6528652170899802,0.97271198886329,0.49787026),
                     c("GO:0006369","termination of RNA polymerase II transcription",0.05573270464712956,-5.525065332784635,1.0639838930348269,4.3543581827353215,-1.3274403722367243,0.9618494424744276,0.28784006),
                     c("GO:0006409","tRNA export from nucleus",0.0180492922400022,-2.9925051872697748,4.6795187705527255,3.864748335629659,-1.5451551399914898,0.9577803481218464,0.57660635),
                     c("GO:0006468","protein phosphorylation",1.0150465485898383,-5.849629271025052,-0.6390268654811149,5.6147158707718114,-2.3400837999301496,0.9426327953530689,0.41776949),
                     c("GO:0006535","cysteine biosynthetic process from serine",0.09469778194088824,-5.889264921377983,1.7744280814969882,4.584579967666666,-1.5001755041604203,0.9361596629318598,0.62197127),
                     c("GO:0006654","phosphatidic acid biosynthetic process",0.11421359458001666,-5.558072357284516,-2.321885604561417,4.665956029453957,-2.4424927980943423,0.9549035373328219,0.56984856),
                     c("GO:0006897","endocytosis",0.6399968963991822,-3.127111856874451,5.13773492441169,5.414408397847148,-1.3876400520322256,0.9691511978736348,0.53769596),
                     c("GO:0006915","apoptotic process",0.3972840732335429,3.1775291496358826,-8.33192955748029,5.2073327066516395,-2.696803942579511,0.9869786354544325,0.00735945),
                     c("GO:0007205","protein kinase C-activating G protein-coupled receptor signaling pathway",0.023705870922346193,1.714670722240622,4.935180714508151,3.9831299247347003,-2.669586226650809,0.855506032915067,0.2027971),
                     c("GO:0007521","muscle cell fate determination",0.00024400927649327026,0.09099161099848761,-7.538882265017623,2,-1.7191940716063332,0.931164622580216,0.51693187),
                     c("GO:0007618","mating",0.00037710524548959944,-3.791368137904168,-6.354960533921752,2.187520720836463,-1.6528652170899802,0.98948114097616,-0),
                     c("GO:0008286","insulin receptor signaling pathway",0.040655889048045374,1.6179390874889665,5.504402504787101,4.217378647939442,-1.7442452133569557,0.8203718285208061,0.65481141),
                     c("GO:0008360","regulation of cell shape",0.6483080002409529,3.0933980119170714,0.9260370495769172,5.420011889394738,-2.5228787452803374,0.7810593250114763,0.53171102),
                     c("GO:0009617","response to bacterium",0.22252413646102945,0.07324387388610921,5.748377069820342,4.9556107920920835,-2.0520763801682738,0.9571119372839897,0.14572441),
                     c("GO:0010045","response to nickel cation",0.0080843477464437,0.5179635044109886,6.607052215776325,3.5160062303860475,-1.5001755041604203,0.9606115104464239,0.67691723),
                     c("GO:0010613","positive regulation of cardiac muscle hypertrophy",0.0017154591559526877,6.089619790358179,-1.3088316508872175,2.8432327780980096,-1.6693833327055616,0.7235420458989957,0.65994664),
                     c("GO:0010629","negative regulation of gene expression",1.1219250764340574,5.413341894082737,2.4903936773260695,5.658193666636185,-1.3114022249188302,0.7867060657482721,0.49623835),
                     c("GO:0010707","globoside biosynthetic process via lactosylceramide",0.00011337804766353971,-5.726826491504109,-2.5366160566771,1.6720978579357175,-1.5451551399914898,0.9648312898576239,0.25078947),
                     c("GO:0010891","negative regulation of sequestering of triglyceride",0.0022010129687726294,5.862755146296142,4.680721847449287,2.951337518795918,-2.8210230527068307,0.8510469041665977,0.45157839),
                     c("GO:0014053","negative regulation of gamma-aminobutyric acid secretion",3.6971102498980334E-05,5.955672848181557,5.039762902357938,1.2041199826559248,-1.5451551399914898,0.8699451468420456,0.59726227),
                     c("GO:0015793","glycerol transmembrane transport",0.03237436208827378,-2.493474489328982,5.171382484259504,4.118463139775366,-1.3882766919926581,0.9690614901535333,0.26740121),
                     c("GO:0021549","cerebellum development",0.009849101705728362,-0.9784330411474381,-6.979849415781065,3.601734148260105,-1.7171513971653551,0.9048578834222591,0.40406595),
                     c("GO:0022614","membrane to membrane docking",0.0004880185529865405,2.402600987777312,-8.414633904366857,2.298853076409707,-1.5001755041604203,0.9963504591346456,0.00501319),
                     c("GO:0030168","platelet activation",0.031036008177810694,-0.1376026097353424,-2.327260974618363,4.100129203359808,-1.3600157519584115,0.7874006443834408,0.38830412),
                     c("GO:0030208","dermatan sulfate biosynthetic process",0.00027605089865905317,-5.975482698167814,1.4052451546093065,2.0530784434834195,-1.5001755041604203,0.9510358910323737,0.3247971),
                     c("GO:0030578","PML body organization",0.0004436532299877641,-2.981308057106973,-1.1588488164930693,2.2576785748691846,-1.4594202834955456,0.9723747790046218,0.25527651),
                     c("GO:0031652","positive regulation of heat generation",0.00176721869945126,5.06724549195376,-3.05830200940739,2.8561244442423,-1.3882766919926581,0.7650495117379952,0.47478216),
                     c("GO:0032060","bleb assembly",0.0011806105398007721,-2.739171314092631,-0.47153695938937684,2.681241237375587,-2.2958494831602008,0.9695555207682965,0.25300851),
                     c("GO:0032076","negative regulation of deoxyribonuclease activity",9.858960666394757E-06,6.304089088656473,4.404633645997789,0.6989700043360189,-1.7977842241988684,0.8472686500727029,0.55436472),
                     c("GO:0032148","activation of protein kinase B activity",0.0001602081108289148,5.899854072067255,-4.3175860136544895,1.8195439355418688,-1.8975662943186637,0.7853929987549769,0.40469454),
                     c("GO:0032487","regulation of Rap protein signal transduction",0.000253868237159665,8.28457139549749,1.8216465653256235,2.0170333392987803,-1.7191940716063332,0.8500502556402696,0.60004102),
                     c("GO:0032570","response to progesterone",0.0020186221964443264,0.7757381813765953,6.844182223298379,2.9138138523837167,-1.4166878480169223,0.9533122398724578,0.55324821),
                     c("GO:0032747","positive regulation of interleukin-23 production",0.0006063260809832775,5.362945664522759,-2.9736098346608926,2.392696953259666,-1.4222784754909792,0.755894531599546,0.54588354),
                     c("GO:0032792","negative regulation of CREB transcription factor activity",2.711214183258558E-05,4.605093883797353,-7.141871589486656,1.0791812460476249,-1.4222784754909792,0.8762208607829973,0.65344236),
                     c("GO:0032879","regulation of localization",0.8977569582819065,2.9857408707679567,3.119404452474719,5.5613888296290535,-1.3658251282373999,0.8664101594255387,0.20748154),
                     c("GO:0032956","regulation of actin cytoskeleton organization",0.3670072050270446,3.5956240165096682,2.836347483279344,5.172906364337767,-1.5377517846450026,0.8259086101059034,0.58714064),
                     c("GO:0032968","positive regulation of transcription elongation by RNA polymerase II",0.05595946074245663,6.127531277237254,-2.687974586452113,4.356121506236986,-1.3392293564723028,0.7614575064126075,0.67754847),
                     c("GO:0035023","regulation of Rho protein signal transduction",0.08740215104775612,7.347598398702844,1.5453433239408847,4.549763225468786,-2.4225082001627745,0.8033322207496058,0.17767807),
                     c("GO:0035821","modulation of process of another organism",0.07254716206366582,-2.4179684572909528,-2.5108446553475683,4.468864040148188,-1.346208812612188,0.9889514253143413,0.65765624),
                     c("GO:0035846","oviduct epithelium development",0.00038696420615599424,-1.4598839542083588,-6.631468387530914,2.1986570869544226,-1.5956795327782693,0.9017176158059406,0.43209483),
                     c("GO:0035984","cellular response to trichostatin A",0.0001306312288297305,1.0052678932165047,6.886557022919951,1.7323937598229686,-1.4594202834955456,0.9518488517936117,0.51480053),
                     c("GO:0036150","phosphatidylserine acyl-chain remodeling",0.0005767491989840933,-5.405853819781651,-2.9128316949553112,2.3710678622717363,-2.728158393463501,0.9634639716669631,0.27035086),
                     c("GO:0036336","dendritic cell migration",0.001158427878301384,-4.337276647344665,-4.86860115388616,2.673020907128896,-1.9677842967020185,0.9657407136531678,0.00522818),
                     c("GO:0038026","reelin-mediated signaling pathway",0.0007887168533115805,2.143073564318831,6.057155912492288,2.506505032404872,-1.3274403722367243,0.8723272139556467,0.37798352),
                     c("GO:0038063","collagen-activated tyrosine kinase receptor signaling pathway",0.00992550865089292,1.9079206648586278,5.397971883886246,3.6050894618815805,-2.500312917381596,0.8548054211255026,0.25301468),
                     c("GO:0039529","RIG-I signaling pathway",0.0005989318604834814,5.951910528886047,-0.42566714342128525,2.387389826338729,-1.390619055749293,0.7240264845566039,0.642558),
                     c("GO:0039534","negative regulation of MDA-5 signaling pathway",0.0004510474504875601,6.9161003716638225,2.4570846529769543,2.2648178230095364,-1.5956795327782693,0.7829740273147665,0.55395978),
                     c("GO:0042262","DNA protection",0.016954947606032382,-0.23930201092876852,5.169065593059701,3.837588438235511,-1.3882766919926581,0.9283110659308078,0.48749122),
                     c("GO:0042981","regulation of apoptotic process",0.6083841390223874,3.113641472992487,2.2790789331518786,5.392408500017495,-1.9062282185012702,0.8476584990235596,0.21341529),
                     c("GO:0043378","positive regulation of CD8-positive, alpha-beta T cell differentiation",0.00018485551249490168,5.379943885552214,-2.0039782208315384,1.8808135922807914,-1.4594202834955456,0.7388262558744074,0.61793949),
                     c("GO:0043418","homocysteine catabolic process",0.008567436819097043,-5.397787171435039,2.758425209627143,3.5412046906832586,-1.6528652170899802,0.9380182020058183,0.07314939),
                     c("GO:0044381","glucose import in response to insulin stimulus",0.0005644254981510998,-0.3772255784679735,6.194426211766302,2.361727836017593,-1.4594202834955456,0.928034936369145,0.69890269),
                     c("GO:0045345","positive regulation of MHC class I biosynthetic process",0.004360125354713081,6.403511870554797,-3.4818447566797954,3.247973266361807,-2.332547047110046,0.7911147512476396,0.5191018),
                     c("GO:0045348","positive regulation of MHC class II biosynthetic process",0.004411884898211653,6.546379319306378,-3.3784889756800274,3.2530955858490316,-1.4222784754909792,0.7910003719070513,0.51952216),
                     c("GO:0045444","fat cell differentiation",0.019557713221960597,-0.15405317374750455,-7.533880929312297,3.8996016591461222,-1.7137681459714469,0.9270962022849388,0.35019661),
                     c("GO:0045603","positive regulation of endothelial cell differentiation",0.0009785018461396797,5.119364218184496,-2.3133338191631316,2.5998830720736876,-2.176525770829699,0.744822729194701,0.56378657),
                     c("GO:0045648","positive regulation of erythrocyte differentiation",0.0027481852857575385,5.449186522661713,-1.7566392841046043,3.04766419460156,-2.8210230527068307,0.7165011874834374,0.4033163),
                     c("GO:0045751","negative regulation of Toll signaling pathway",0.00026372719782605974,6.919042657714815,3.026106963202875,2.03342375548695,-1.353988590408761,0.8226780651457645,0.39482722),
                     c("GO:0045917","positive regulation of complement activation",0.00033273992249082307,7.4372664057539755,-1.2682960224566364,2.1335389083702174,-1.7191940716063332,0.7712152997791639,0.44150672),
                     c("GO:0045944","positive regulation of transcription by RNA polymerase II",1.0624558256943641,5.769606416214144,-1.7353859520264632,5.63454074708158,-4,0.717903316081886,0),
                     c("GO:0046326","positive regulation of glucose import",0.004715047938703292,6.9435438524360995,-2.9369558345849693,3.281941933440825,-3.2596373105057563,0.7906372380226981,0.41832432),
                     c("GO:0046339","diacylglycerol metabolic process",0.028526902688213226,-5.443816769268588,-3.2492493919978367,4.063520999689991,-2.623423042943488,0.9682347668268916,0.42084756),
                     c("GO:0046544","development of secondary male sexual characteristics",0.0002045734338276912,-1.728099560093911,-6.6736776025192475,1.9242792860618816,-1.5956795327782693,0.9184159542089377,0.35257589),
                     c("GO:0046551","retinal cone cell fate commitment",7.394220499796068E-06,-0.7813678732059797,-7.275051786230126,0.6020599913279624,-1.4594202834955456,0.9185083020479073,0.30376398),
                     c("GO:0046626","regulation of insulin receptor signaling pathway",0.017891548869339885,7.705851199230559,1.6900306052071465,3.8609366207000937,-1.8948305720006684,0.8241488867122213,0.50107779),
                     c("GO:0046628","positive regulation of insulin receptor signaling pathway",0.004761878001868667,6.917057171738043,-1.116177340223685,3.286231854028553,-1.837435593476981,0.7555543579226078,0.46204972),
                     c("GO:0046777","protein autophosphorylation",0.001434478776960437,-5.9131989289992415,-1.375535870049579,2.765668554759014,-1.6145724851948695,0.9601126234036965,0.28031619),
                     c("GO:0046834","lipid phosphorylation",0.015197587867247515,-5.706408881767255,-2.8714482928032927,3.7900739484263046,-3.167491087293764,0.9650663971038465,0.00493514),
                     c("GO:0048026","positive regulation of mRNA splicing, via spliceosome",0.034464461749549466,6.138881468056768,-2.931797810748323,4.145631415290328,-1.7817271464285525,0.763956147411112,0.62701823),
                     c("GO:0048311","mitochondrion distribution",0.019658767568791147,-2.977657985433961,2.0708121129665438,3.9018395920512297,-1.4677553563734178,0.946487297370937,0.59952363),
                     c("GO:0050729","positive regulation of inflammatory response",0.013491987671961225,7.028120861582177,-1.2117237942874135,3.738384123512156,-1.666149854897455,0.7394294343267176,0.484979),
                     c("GO:0050821","protein stabilization",0.10143638155636905,0.7505258557063659,1.0130778682932366,4.614433158550452,-1.5403060235220296,0.8571771039168861,0.54859689),
                     c("GO:0050901","leukocyte tethering or rolling",0.004922086112697582,-4.34915598161059,-4.9555692342506585,3.3005954838899636,-1.7953374882517812,0.9558537450191036,0.69904619),
                     c("GO:0051045","negative regulation of membrane protein ectodomain proteolysis",0.007867450611783015,5.843687736253752,3.503552938298336,3.504198918539445,-3.552841968657781,0.8126574638647425,0.24434509),
                     c("GO:0051085","chaperone cofactor-dependent protein refolding",0.17454303963785278,-5.7957101155844475,0.23424039237859007,4.850137524920791,-1.4677553563734178,0.9565943329272775,0.14432087),
                     c("GO:0051290","protein heterotetramerization",0.0013851839736284633,-2.7765857959494022,0.014092954305155845,2.7505083948513462,-1.9978339382434924,0.9682021087696818,0.2684544),
                     c("GO:0051638","barbed-end actin filament uncapping",8.133642549775675E-05,3.4255834974692445,-1.4234058057269658,1.5314789170422551,-1.7977842241988684,0.7715231441181621,0.45946467),
                     c("GO:0051897","positive regulation of phosphatidylinositol 3-kinase/protein kinase B signal transduction",0.026927286320090678,6.670711096192131,-1.022248221070941,4.038461196178564,-1.3029451077274257,0.7322672247803768,0.59526557),
                     c("GO:0055005","ventricular cardiac myofibril assembly",0.00034999310365701385,-1.3377591684899024,-5.340092871610168,2.155336037465062,-3.552841968657781,0.8692834228862545,-0),
                     c("GO:0055129","L-proline biosynthetic process",0.09290838057993758,-5.687949679158526,1.5725561675622197,4.576295268770989,-1.3882766919926581,0.9456023778400062,0.38107409),
                     c("GO:0060025","regulation of synaptic activity",0.000253868237159665,1.3618271292596356,1.506212135944049,2.0170333392987803,-1.5001755041604203,0.8542345085221912,0.68303869),
                     c("GO:0060029","convergent extension involved in organogenesis",0.0009686428854732849,-0.4171907343193511,-7.309178314268338,2.595496221825574,-1.3567447749752284,0.9094093254394764,0.64368634),
                     c("GO:0060047","heart contraction",0.050813083274598575,-1.7535318047501887,-6.015623389950335,4.314225470926496,-2.5086383061657274,0.9316935268545218,0.30186478),
                     c("GO:0060612","adipose tissue development",0.00502067571936153,-0.5424268976462351,-7.390229618702706,3.3092041796704077,-1.3734543409728706,0.9122933178624216,0.4603377),
                     c("GO:0060770","negative regulation of epithelial cell proliferation involved in prostate gland development",0.000650691403982054,5.001516775138499,1.353121759813428,2.423245873936808,-1.3882766919926581,0.7771630443947797,0.56804024),
                     c("GO:0060936","cardiac fibroblast cell development",3.4506362332381646E-05,-0.3333432650348204,-6.737143958704935,1.1760912590556813,-1.5956795327782693,0.9130495675105116,0.61805987),
                     c("GO:0061076","negative regulation of neural retina development",8.87306459975528E-05,5.095249095941459,1.9055402328449023,1.568201724066995,-1.7977842241988684,0.8045036896298703,0.10394042),
                     c("GO:0061083","regulation of protein refolding",4.190058283217772E-05,2.8223756519676897,-7.971771590797005,1.255272505103306,-2.1958605676646497,0.9007934769858642,0.1070597),
                     c("GO:0061337","cardiac conduction",0.02085170180942491,2.5133182493846187,-2.377396306335231,3.927421695050419,-2.2479515521805613,0.7952728699116043,0.47843103),
                     c("GO:0061458","reproductive system development",0.1404162472911273,-1.119045002454251,-6.777147037141997,4.75565384260343,-1.3298469548078198,0.9035382413573521,0.47054433),
                     c("GO:0061635","regulation of protein complex stability",0.0017622892191180627,-0.07333497583772439,0.6084745361365091,2.8549130223078554,-1.3567447749752284,0.8833888765786825,0.30006872),
                     c("GO:0061738","late endosomal microautophagy",0.0027950153489229132,-2.2385869273406986,6.40947501287549,3.0549958615291417,-1.3882766919926581,0.9702715742367721,0.5211141),
                     c("GO:0061740","protein targeting to lysosome involved in chaperone-mediated autophagy",0.0035837322022344944,-3.441747128025423,4.085413619288951,3.162862993321926,-1.5956795327782693,0.9203404684531251,0.34828093),
                     c("GO:0070094","positive regulation of glucagon secretion",0.00010844856733034232,5.333455520056423,-1.0213547088477763,1.6532125137753437,-1.7977842241988684,0.78003473954788,0.52397596),
                     c("GO:0070168","negative regulation of biomineral tissue development",0.006595644685818093,5.008358382785751,1.6951350920438204,3.4276483711869328,-1.390619055749293,0.7618641889900627,0.56345285),
                     c("GO:0070358","actin polymerization-dependent cell motility",0.0019151031094471816,-4.7557596707540855,-5.021598366237832,2.890979596989689,-1.4222784754909792,0.9757051251941771,0.42384923),
                     c("GO:0070375","ERK5 cascade",0.00012816648866313183,2.6005436864895524,6.6460054453599104,1.724275869600789,-2.6861327796308467,0.886166933955363,0.11366838),
                     c("GO:0070459","prolactin secretion",0.0019052441487807867,-0.678538426775829,3.3327324986043205,2.8887409606828927,-1.4222784754909792,0.8334306374927862,0.26335126),
                     c("GO:0070474","positive regulation of uterine smooth muscle contraction",0.0001626728509955135,4.908701313463943,-3.4583356149066398,1.8260748027008264,-1.6528652170899802,0.7777545798076938,0.58864769),
                     c("GO:0070668","positive regulation of mast cell proliferation",0.0004214705684883758,6.6890935819579855,-4.073055293902425,2.2355284469075487,-1.5001755041604203,0.8142228787804305,0.48154824),
                     c("GO:0070814","hydrogen sulfide biosynthetic process",0.03377679924306844,-5.813398761480927,2.371601143854269,4.1368790398755175,-1.4222784754909792,0.959771249457285,0.41474247),
                     c("GO:0071277","cellular response to calcium ion",0.06043542888499986,0.8872311570315411,6.466265030396026,4.389538177331949,-1.7113037394097443,0.9480380778209634,0.37749893),
                     c("GO:0071673","positive regulation of smooth muscle cell chemotaxis",0.00021443239449408597,7.44578983579099,-1.4778750593670533,1.9444826721501687,-1.5451551399914898,0.7802513508252981,0.58540694),
                     c("GO:0071864","positive regulation of cell proliferation in bone marrow",0.0003475283634904152,6.58941759864256,-4.2012780597654285,2.1522883443830563,-3.8239087409443187,0.8170839550203803,0.35458562),
                     c("GO:0071932","replication fork reversal",0.007322743034964705,-5.180865523646666,3.488435053126829,3.4730488050885375,-1.3567447749752284,0.9721250340885588,0.06646365),
                     c("GO:0072177","mesonephric duct development",0.0007739284123119885,-0.9260419696411317,-6.561681637670804,2.4983105537896004,-1.5956795327782693,0.8971174803681708,0.6722196),
                     c("GO:0072287","metanephric distal tubule morphogenesis",7.394220499796067E-05,-0.7982684810147064,-6.566738522387712,1.4913616938342726,-1.7191940716063332,0.8957219369314681,0.66294094),
                     c("GO:0072659","protein localization to plasma membrane",0.1961243045365909,-3.1849148986337728,4.395608774382799,4.9007657318060245,-1.356942316248547,0.9602724969753796,0.37156993),
                     c("GO:0086047","membrane depolarization during Purkinje myocyte cell action potential",0.00023661505599347418,1.9930787955103257,-0.39689434847555666,1.9867717342662448,-1.3882766919926581,0.7989154549412761,0.68960777),
                     c("GO:0090083","regulation of inclusion body assembly",0.0026914962619257687,4.193508567542767,5.783316284188521,3.038620161949703,-1.46167766676856,0.8692702864364436,0.50466917),
                     c("GO:0090204","protein localization to nuclear pore",0.00015034915016252004,-3.9271420165699626,4.468708237238987,1.792391689498254,-3.2596373105057563,0.9708900354812,0.00405798),
                     c("GO:0097212","lysosomal membrane organization",0.00047323011198694835,-3.3575671943227476,-0.8622676497233106,2.285557309007774,-1.4594202834955456,0.9723004269639097,0.25611759),
                     c("GO:0097241","hematopoietic stem cell migration to bone marrow",0.0004584416709873562,-4.639806535346016,-5.1996230559370415,2.271841606536499,-1.5956795327782693,0.9750459809114913,0.44723519),
                     c("GO:0097278","complement-dependent cytotoxicity",0.00019717921332789513,-3.380643354497799,-3.612970907386261,1.9084850188786497,-1.3882766919926581,0.9964936933458816,0.0041008),
                     c("GO:0098792","xenophagy",0.0014788440999592134,-1.0557345852687245,5.64782622431738,2.7788744720027396,-1.6354490046460282,0.9377565470428497,0.53218536),
                     c("GO:0099022","vesicle tethering",0.028662463397376152,-3.0622297050527183,3.197176317638386,4.0655797147284485,-1.5322439487559671,0.9366057504416032,0.34171025),
                     c("GO:0099175","regulation of postsynapse organization",0.027792410118566816,1.7410304026146406,1.820818672691383,4.0521935804334746,-2.7825160557860937,0.8284086441869107,0.45961659),
                     c("GO:0106227","peptidyl-lysine glutarylation",7.394220499796068E-06,-6.319545110608982,-1.130045646124264,0.6020599913279624,-1.7191940716063332,0.9742468724770326,0.03408306),
                     c("GO:0120323","lipid ubiquitination",1.2323700832993446E-05,-5.599890945320626,-3.528062449743901,0.7781512503836436,-2.0920514783877278,0.9779342767851821,0.39185285),
                     c("GO:0140042","lipid droplet formation",0.012644117054651276,-3.0203348031456514,2.512332033397488,3.7102020146553847,-1.6029294500405913,0.9387832213991439,0.45439802),
                     c("GO:0150062","complement-mediated synapse pruning",0.00022429135516048072,-2.966030668330845,0.9278678518974907,1.9637878273455553,-1.4594202834955456,0.9733284259143316,0.43331549),
                     c("GO:0150064","vertebrate eye-specific patterning",0.0001774612919951056,-1.237665352618459,-7.098010709535665,1.863322860120456,-1.4594202834955456,0.9211016902302818,0.51012229),
                     c("GO:1900029","positive regulation of ruffle assembly",0.007985758139779753,6.25262323373837,-2.1090912151752157,3.51067903103221,-3.638272163982407,0.7715903570427525,0.43409096),
                     c("GO:1900076","regulation of cellular response to insulin stimulus",0.019934818467450197,7.929455531869915,1.7696627857137053,3.907894835416283,-1.7070797003999938,0.8311224785052777,0.46341696),
                     c("GO:1900223","positive regulation of amyloid-beta clearance",0.0010031492478056664,5.11222572759569,-3.2645681851351065,2.61066016308988,-1.5001755041604203,0.7706322823143,0.41247998),
                     c("GO:1901032","negative regulation of response to reactive oxygen species",1.9717921332789515E-05,7.293277469800174,3.1493510596042795,0.9542425094393249,-2.0338582672609675,0.8365519679273687,0.32080048),
                     c("GO:1901165","positive regulation of trophoblast cell migration",0.0006457619236488565,5.1857714039107385,-2.4559447178374865,2.419955748489758,-1.3567447749752284,0.7522279722589473,0.49740704),
                     c("GO:1902038","positive regulation of hematopoietic stem cell differentiation",0.00022182661499388205,6.7704823981733355,-2.596591972375019,1.9590413923210936,-1.7191940716063332,0.7889674593478646,0.52821246),
                     c("GO:1902527","positive regulation of protein monoubiquitination",0.0003918936864891916,6.0856988001280365,-4.201692715036914,2.2041199826559246,-1.5956795327782693,0.8029372185833991,0.47639316),
                     c("GO:1902870","negative regulation of amacrine cell differentiation",0.000325345701991027,5.550636757102981,4.0519580725012005,2.123851640967086,-1.7977842241988684,0.8216152392600102,0.46906492),
                     c("GO:1903347","negative regulation of bicellular tight junction assembly",0.0003992879069889876,5.134495302575139,4.899696428977196,2.2121876044039577,-1.7977842241988684,0.8460332811509855,0.46062898),
                     c("GO:1903613","regulation of protein tyrosine phosphatase activity",1.2323700832993446E-05,4.793611450419846,-6.912810868122157,0.7781512503836436,-1.5451551399914898,0.8668647350319719,0.63970968),
                     c("GO:1903781","positive regulation of cardiac conduction",0.00010105434683054626,4.863089772462669,-3.6329900718512484,1.6232492903979006,-1.4222784754909792,0.7838816043038966,0.6646122),
                     c("GO:1903860","negative regulation of dendrite extension",0.00014788440999592134,5.0997893836881,3.6996096428915526,1.7853298350107671,-1.3274403722367243,0.8178227504021721,0.53886376),
                     c("GO:1903949","positive regulation of atrial cardiac muscle cell action potential",7.394220499796068E-06,3.9583897266547594,-2.226555274385862,0.6020599913279624,-1.5001755041604203,0.7821748725057474,0.59360644),
                     c("GO:1904450","positive regulation of aspartate secretion",1.4788440999592136E-05,7.29756404482241,-3.695485301042683,0.8450980400142568,-1.7191940716063332,0.8310297436242544,0.48333657),
                     c("GO:1904566","cellular response to 1-oleoyl-sn-glycerol 3-phosphate",7.394220499796068E-06,1.265082335106395,7.071138000223974,0.6020599913279624,-1.7977842241988684,0.9599918531276044,0.28547526),
                     c("GO:1904613","cellular response to 2,3,7,8-tetrachlorodibenzodioxine",0.0013679307924622725,1.0732972655337685,6.73079257028051,2.7450747915820575,-1.7191940716063332,0.9511651134990943,0.31676361),
                     c("GO:1904715","negative regulation of chaperone-mediated autophagy",0.00017253181166190822,6.190488849920306,4.02688129546415,1.8512583487190752,-1.4594202834955456,0.8409399053731395,0.69325143),
                     c("GO:1904764","chaperone-mediated autophagy translocation complex disassembly",0.0025879771749286237,-2.9042873275406786,0.5992937119816611,3.021602716028242,-1.7977842241988684,0.9695280634206972,0.28793683),
                     c("GO:1905203","regulation of connective tissue replacement",0.00032781044215762567,2.49090169056875,-3.8467574242196685,2.1271047983648077,-1.6528652170899802,0.8416961481704126,0.34180262),
                     c("GO:1905383","protein localization to presynapse",0.0003549225839902112,-3.6953751855656884,4.360333469432143,2.161368002234975,-2.1958605676646497,0.9707128874389409,0.23577969),
                     c("GO:1990519","pyrimidine nucleotide import into mitochondrion",0.010280431234883133,-2.711213326919617,5.384792861635989,3.6203442997544935,-1.6528652170899802,0.9696021852085271,0.11520004),
                     c("GO:1990743","protein sialylation",1.9717921332789515E-05,-6.308608192514768,-0.804556564143887,0.9542425094393249,-1.3274403722367243,0.9732515822910539,0.19780103),
                     c("GO:1990832","slow axonal transport",0.0025879771749286237,-3.5448597358985965,4.801056055685872,3.021602716028242,-1.7977842241988684,0.968703080536734,0.2223456),
                     c("GO:1990839","response to endothelin",4.43653229987764E-05,0.4223132578288504,7.003298263701399,1.2787536009528289,-3.8239087409443187,0.9608275602242001,-0),
                     c("GO:2000051","negative regulation of non-canonical Wnt signaling pathway",0.0065512793628193155,6.585019282920462,2.8000334324193568,3.424718337331567,-1.8233300673318504,0.7962786291479366,0.47088909),
                     c("GO:2000111","positive regulation of macrophage apoptotic process",0.00010598382716374363,6.121978180961389,-4.666228316618728,1.6434526764861874,-1.5451551399914898,0.823166471944115,0.5344573),
                     c("GO:2000259","positive regulation of protein activation cascade",0.0035230893243854726,6.157888843435619,-3.5071125052084406,2.9800033715837464,-1.7191940716063332,0.7774900340216702,0.51162432),
                     c("GO:2000340","positive regulation of chemokine (C-X-C motif) ligand 1 production",0.00038203472582279684,5.38737703863321,-3.1338630794986724,2.1931245983544616,-1.5956795327782693,0.7604241170215129,0.4448422),
                     c("GO:2000427","positive regulation of apoptotic cell clearance",0.0004461179701543627,6.89852700401224,-2.2226643910768464,2.2600713879850747,-1.5956795327782693,0.7889893114877199,0.55723394),
                     c("GO:2000987","positive regulation of behavioral fear response",0.00024400927649327026,6.411726663000238,-1.153854395009761,2,-1.5451551399914898,0.7344753397531082,0.61613263),
                     c("GO:2001016","positive regulation of skeletal muscle cell differentiation",0.0004855538128199418,6.54747334761751,-2.459994205527089,2.296665190261531,-1.3274403722367243,0.782199003096027,0.5464115));


BN.snp.revigo.names <- c("term_ID","description","frequency","plot_X","plot_Y","log_size","value","uniqueness","dispensability");
BN.snp.revigo.data <- rbind(c("GO:0000381","regulation of alternative mRNA splicing, via spliceosome",0.09320661414009604,2.482291886447794,-0.6687676666077658,4.577687073553896,-2.0856568428805593,0.846367182549155,0.34115005),
                     c("GO:0001568","blood vessel development",0.15784688974931324,-4.174111731976727,-4.614987630284883,4.806471667609058,-2.7351821769904636,0.8203696064622389,0.67328265),
                     c("GO:0001578","microtubule bundle formation",0.14962205181337343,-5.430832385445306,3.1320954823300258,4.783231617568076,-3.0555173278498313,0.9526360310265082,0.00795379),
                     c("GO:0001660","fever generation",0.001547856824623977,-3.333869176126269,-1.7951147935245908,2.798650645445269,-1.382370702242158,0.8645290819593562,0.5528132),
                     c("GO:0001701","in utero embryonic development",0.03255921760076868,-4.352678484162369,-4.506639472695552,4.120935692561131,-2.9913998282380825,0.850630003237518,0.39618226),
                     c("GO:0001702","gastrulation with mouth forming second",0.005087223703859695,-4.459703690503492,-4.630860148077866,3.31492005599242,-1.817585347565446,0.857400119746779,0.62546471),
                     c("GO:0001933","negative regulation of protein phosphorylation",0.026256876994775837,6.1435248252523005,0.3889255129318057,4.027512692448811,-2.769551078621726,0.7689208936670073,0.6121628),
                     c("GO:0002021","response to dietary excess",0.0028319864514218937,-3.103617354469405,-1.5151891668648545,3.060697840353612,-1.6115435472997333,0.8527729461186297,0.64073523),
                     c("GO:0002244","hematopoietic progenitor cell differentiation",0.021172118031082737,-3.5958138840127263,-5.979481381367678,3.9340437190355253,-1.792904459580782,0.8842888834347774,0.58076808),
                     c("GO:0002674","negative regulation of acute inflammatory response",0.0013999724146280554,6.903087974865367,1.7262520189216408,2.7551122663950713,-1.8975662943186637,0.8016481999750805,0.52199865),
                     c("GO:0002792","negative regulation of peptide secretion",0.009462137499572367,7.258316547825423,-0.2884990440972621,3.584331224367531,-2.2328441339178196,0.7808849750945946,0.59128304),
                     c("GO:0002934","desmosome organization",0.0026471309389269923,-5.883395352744218,2.1969114363678988,3.031408464251624,-1.3019864960608183,0.9663783204893004,0.54806111),
                     c("GO:0003013","circulatory system process",0.16877061816767863,-4.904532604843416,-3.6763291287732582,4.835532040925999,-1.392223396258307,0.8822665620863299,0.61568661),
                     c("GO:0003192","mitral valve formation",0.00026372719782605974,-4.244427464786994,-5.14276744913542,2.03342375548695,-1.917933065714887,0.8499550132944855,0.53059195),
                     c("GO:0006089","lactate metabolic process",0.15974966915792743,-7.09139256891967,1.9753466179608554,4.811675525366911,-1.8975662943186637,0.9853164478772916,0.54429),
                     c("GO:0006278","RNA-templated DNA biosynthetic process",0.1788218285670681,-2.8683930218214475,5.937140679737178,4.860655374823497,-1.678194516142461,0.9762911695095656,0.4490849),
                     c("GO:0006313","DNA transposition",0.31263750169204413,-3.2912014283670494,5.563815019559114,5.103273352793802,-1.9673812391492802,0.9792124564908667,0.23320655),
                     c("GO:0006325","chromatin organization",1.3384303174082528,-5.027174211153823,3.024290297548671,5.734825422616354,-1.9978339382434924,0.9526243542743458,0.54950096),
                     c("GO:0006376","mRNA splice site recognition",0.056284806444447666,-4.399265892255172,3.955925430598413,4.358639051893465,-1.3475602524105799,0.9384319255796355,0.47843915),
                     c("GO:0006670","sphingosine metabolic process",0.05907242757287078,-3.979357444335574,5.932920060333985,4.379631796019372,-3.0757207139381184,0.9706899660613872,0.00743248),
                     c("GO:0006686","sphingomyelin biosynthetic process",0.005592495438012425,-3.2639397354142283,6.276853637685023,3.3560258571931225,-1.3019864960608183,0.9721186037805051,0.6351172),
                     c("GO:0006869","lipid transport",0.9456493244590856,-7.161076576363113,-0.4209849090502218,5.5839601059765185,-1.7413627172759236,0.97919862848792,0.57273937),
                     c("GO:0006897","endocytosis",0.6399968963991822,-7.127884009816407,-0.9972810455385739,5.414408397847148,-1.489454989793388,0.9712013943266251,0.66216709),
                     c("GO:0006913","nucleocytoplasmic transport",0.5732591269081895,-7.011802767754504,-0.6187334814380726,5.3665817025364975,-1.362210170637771,0.9720940991224037,0.46768014),
                     c("GO:0006971","hypotonic response",0.0013753250129620686,0.5434557204613188,4.929538253335258,2.747411807886423,-1.9186526921958675,0.9464547970170871,0.40232125),
                     c("GO:0006979","response to oxidative stress",0.8191810417707402,1.4767482423495046,5.248410921963154,5.521610056993851,-2.042392712939905,0.9342584717056688,0.44442572),
                     c("GO:0007000","nucleolus organization",0.013161712489636998,-5.781279140216097,3.0077309280444315,3.727622577969137,-2.166215625343521,0.963271660782368,0.32491698),
                     c("GO:0007076","mitotic chromosome condensation",0.03710173372781007,-5.437601365493472,2.687269538740475,4.177651911698254,-3.0268721464003012,0.9523717741870281,0.34829427),
                     c("GO:0007229","integrin-mediated signaling pathway",0.0840648928621815,3.102409163819955,3.2030287023363355,4.5328562543112625,-2.9546770212133424,0.8071491916458137,0.53861329),
                     c("GO:0007259","receptor signaling pathway via JAK-STAT",0.020538679808266878,3.1434777333536137,3.8312431708823125,3.9208534961212593,-1.8999742698921374,0.8220131421466671,0.49153737),
                     c("GO:0007522","visceral muscle development",0.00022429135516048072,-3.4180339154764807,-6.212780350844708,1.9637878273455553,-3.744727494896694,0.8874581792180969,-0),
                     c("GO:0007612","learning",0.012193069604163715,-4.961866884607586,-3.494301623714495,3.694429690957083,-1.4046140191908583,0.8960798396329288,0.63668398),
                     c("GO:0008057","eye pigment granule organization",0.00032288096182442826,-5.820421690292575,3.495839167716952,2.12057393120585,-1.8184422261372137,0.9682758836739628,0.45289087),
                     c("GO:0008219","cell death",0.4651211168388386,-7.229695581702569,1.1335103034231802,5.275797216010534,-1.8127613801685214,0.9953726677909294,0.0090814),
                     c("GO:0009048","dosage compensation by inactivation of X chromosome",0.0023020673156031756,1.3932058290591063,1.111899389485566,2.9708116108725178,-3.0861861476162833,0.7871459375531901,0.47787801),
                     c("GO:0009056","catabolic process",9.788032838620547,-7.284794146700128,0.7060251476491619,6.598924384475181,-1.392974121565214,0.9940581149254236,0.04794448),
                     c("GO:0009268","response to pH",0.012496232644655355,0.8540588092836636,5.561272722643507,3.7050936105478733,-1.4028535121663046,0.9467205082583371,0.53065188),
                     c("GO:0009299","mRNA transcription",0.011650826767512005,-2.875657246192721,5.66039167655615,3.674677467873199,-1.9730583720409707,0.9772173530634107,0.17215736),
                     c("GO:0009410","response to xenobiotic stimulus",0.16588440743259156,2.283613464103183,5.897653771500228,4.828040875907554,-2.474955192963155,0.9250464755201185,0.3759818),
                     c("GO:0010042","response to manganese ion",0.003958372707557495,3.053973697077737,6.309322384355203,3.2060158767633444,-1.4869156395348557,0.9340180260948427,0.65539321),
                     c("GO:0010575","positive regulation of vascular endothelial growth factor production",0.006901272466476329,4.81438690067782,-3.4244324705450047,3.4473131088235682,-1.7770235501066087,0.746435080674424,0.63985713),
                     c("GO:0010594","regulation of endothelial cell migration",0.03429439467805416,6.593253529961348,-2.2780754240836316,4.143483210670062,-1.8210230527068305,0.7897023721547208,0.59242718),
                     c("GO:0010616","negative regulation of cardiac muscle adaptation",0.0003130220011580335,6.85990852023714,-0.07608613534155337,2.1072099696478683,-1.5055671012736014,0.7600028757194274,0.57264968),
                     c("GO:0010628","positive regulation of gene expression",0.5423685384002082,4.375374517758638,-2.707818621867504,5.342525320115581,-3.6989700043360187,0.75197175355301,0.39342096),
                     c("GO:0010763","positive regulation of fibroblast migration",0.0010869504134700218,5.449340938836008,-5.050267203168399,2.645422269349092,-1.5018273393634562,0.8292249124914274,0.60063743),
                     c("GO:0010770","positive regulation of cell morphogenesis",0.007966040218446963,3.6433985622670626,-4.496456290720043,3.5096057046115563,-3.3767507096020997,0.7608858951385074,0.41410062),
                     c("GO:0010976","positive regulation of neuron projection development",0.04195727185600948,4.3503963370623655,-3.647861154591056,4.231061610614955,-2.4546928835341757,0.7731826953112156,0.59929015),
                     c("GO:0014031","mesenchymal cell development",0.0006063260809832775,-3.354896979919642,-5.8582920739191255,2.392696953259666,-1.66334017654558,0.8807888766691475,0.5891425),
                     c("GO:0014898","cardiac muscle hypertrophy in response to stress",0.001252088004632134,-3.265362539839178,-2.1742882936283174,2.7067177823367587,-2.6270879970298933,0.8643840929974207,0.6854827),
                     c("GO:0016055","Wnt signaling pathway",0.18316963022094818,3.191767750359589,2.893761391900587,4.8710881699072415,-1.9125735429637145,0.7953640697765477,0.59217588),
                     c("GO:0016202","regulation of striated muscle tissue development",0.0017228533764524837,0.8616834297510638,-6.705089424146391,2.845098040014257,-2.6345120151091,0.8434214277692719,0.11385893),
                     c("GO:0016485","protein processing",0.5034847975319472,-2.3116989231127967,6.281165838110699,5.310217345194196,-1.5590909179347823,0.9715048212761567,0.61324128),
                     c("GO:0016567","protein ubiquitination",1.2678648064385323,-1.9236258270083755,6.696115627089043,5.711302648460635,-1.7617029321246063,0.9744264854985575,0.40604291),
                     c("GO:0018243","protein O-linked glycosylation via threonine",0.0049418040340303716,-1.7770273351092603,6.798587967643679,3.302330928684399,-1.325322532126801,0.9746086329747766,0.44120675),
                     c("GO:0018279","protein N-linked glycosylation via asparagine",0.04641598681738651,-2.0291538014539956,6.6667154573602145,4.274919506403754,-1.3175841383226414,0.972490778629794,0.55171726),
                     c("GO:0018401","peptidyl-proline hydroxylation to 4-hydroxy-L-proline",2.4647401665986892E-05,-1.7398290705303596,7.024423227785193,1.0413926851582251,-1.9500071430798573,0.9790330226267453,0.34342236),
                     c("GO:0019082","viral protein processing",0.000768998931978791,-2.1137007800228935,3.2045944274506177,2.4955443375464483,-1.7999707334462298,1,-0),
                     c("GO:0019471","4-hydroxyproline metabolic process",0.0001306312288297305,-4.447144665853498,5.447736631004345,1.7323937598229686,-1.8474059220725303,0.9847633726867191,0.14718643),
                     c("GO:0021515","cell differentiation in spinal cord",0.01231877135266025,-4.166311477135258,-4.458853140653567,3.69888313675259,-1.6897316333675525,0.8561149240844089,0.595661),
                     c("GO:0021591","ventricular system development",0.003573873241568099,-4.234341884512736,-4.771532059740936,3.161667412437736,-3.070581074285707,0.8571065930641497,0.4332494),
                     c("GO:0030010","establishment of cell polarity",0.11765190711242182,-1.0180133876356656,-8.112374757105147,4.678836926900922,-3.2676062401770314,0.9928325481170667,-0),
                     c("GO:0030032","lamellipodium assembly",0.019353139788132908,-4.800852472392426,2.606649907020678,3.8950355974523228,-1.4822764051662645,0.9603887483288079,0.49268407),
                     c("GO:0030070","insulin processing",0.0006728740654814422,-1.5350218619324791,5.780194303408393,2.437750562820388,-2.9355420107730814,0.876936190113164,0.10214819),
                     c("GO:0030111","regulation of Wnt signaling pathway",0.2736946070597848,4.958327206772773,0.42708943164820523,5.045499008437883,-4,0.7984636943076198,0),
                     c("GO:0030220","platelet formation",0.0023735447804345373,-3.597107950065385,-6.077782967054742,2.984077033902831,-1.7822529267372063,0.8806905534751287,0.66001329),
                     c("GO:0030278","regulation of ossification",0.021842527356397586,7.26611486029314,-2.1977340907341425,3.9475807493043225,-1.6212388246836267,0.8095901605790968,0.50675978),
                     c("GO:0030279","negative regulation of ossification",0.00778364944611866,7.022619281083362,-0.23906532339474296,3.499549625905149,-1.7778039536982801,0.762935195311119,0.5999615),
                     c("GO:0030308","negative regulation of cell growth",0.08751552909541965,5.9566131318967575,-0.2272546187494185,4.550326211208258,-3.853871964321762,0.7602060006473594,0.5153367),
                     c("GO:0030324","lung development",0.025357246833967312,-4.027190536936834,-4.75267196906731,4.012373167222489,-1.8774564759312458,0.8429464234166962,0.47948347),
                     c("GO:0030382","sperm mitochondrion organization",0.0005028069939861325,-6.09876395482547,2.9518964098859173,2.311753861055754,-1.5469876088785448,0.9688315414014779,0.26820528),
                     c("GO:0030513","positive regulation of BMP signaling pathway",0.012547992188153927,5.397784507488206,-2.7785267502786724,3.706888394981618,-2.1817741063860443,0.7707167039233039,0.62303062),
                     c("GO:0030593","neutrophil chemotaxis",0.013171571450303394,1.3405865968998263,4.8546228278733015,3.727947709544797,-1.3692671071828035,0.8939302662674433,0.56153576),
                     c("GO:0030644","intracellular chloride ion homeostasis",0.0007443515303128041,-5.080443737923308,-0.0744240807306197,2.481442628502305,-2.5346171485515816,0.9926030552600298,-0),
                     c("GO:0030728","ovulation",0.0021270707637746685,-5.275757940353852,-3.7577416904046093,2.936513742478893,-1.3262420634504233,0.9094049923826121,0.54207276),
                     c("GO:0030878","thyroid gland development",0.0031277552714137364,-4.358419555402439,-4.896259114130429,3.103803720955957,-1.8632794328435933,0.8590201340547419,0.43042516),
                     c("GO:0030948","negative regulation of vascular endothelial growth factor receptor signaling pathway",0.0008947006804753241,6.84862624068217,1.5838643603157445,2.561101383649056,-2.0942041196321317,0.8009391616285847,0.65072455),
                     c("GO:0031032","actomyosin structure organization",0.14377568813820132,-5.510603379070177,3.0522047611278897,4.7659217573037775,-1.3774751375964318,0.9471605279607451,0.58908222),
                     c("GO:0031547","brain-derived neurotrophic factor receptor signaling pathway",0.002055593298943307,3.538527794148614,4.720690503519712,2.921686475483602,-2.2984320149440727,0.8388914266145914,0.60505026),
                     c("GO:0031669","cellular response to nutrient levels",0.3309948864528712,2.0264068643079653,5.116599029020445,5.128053375731982,-2.2034256667895704,0.9128061965649626,0.42983343),
                     c("GO:0032014","positive regulation of ARF protein signal transduction",0.00012570174849653313,6.11187422045942,-3.493215161741329,1.7160033436347992,-1.4555598626823076,0.8120356537904863,0.40294129),
                     c("GO:0032691","negative regulation of interleukin-1 beta production",0.004318224771880904,6.433653890432107,-0.5693236064185891,3.243781916093795,-3.5086383061657274,0.7491077763603409,0.49899857),
                     c("GO:0032753","positive regulation of interleukin-4 production",0.0020358753776105174,4.967647742858971,-3.700038601943734,2.9175055095525466,-1.3175841383226414,0.760688087301741,0.65582536),
                     c("GO:0032879","regulation of localization",0.8977569582819065,3.563444038411494,0.2383626648825884,5.5613888296290535,-3.309803919971486,0.8514336601505229,0.21592382),
                     c("GO:0032922","circadian regulation of gene expression",0.061155133013646674,2.052291373426742,-0.5733780855741074,4.394679275545099,-1.9079815292472029,0.8496001618429241,0.29304552),
                     c("GO:0032944","regulation of mononuclear cell proliferation",0.06872927954560445,4.964524506716173,1.899240071474628,4.445386223032164,-1.9783972839717578,0.8444525907624233,0.68631809),
                     c("GO:0032966","negative regulation of collagen biosynthetic process",0.00028098037899225057,7.656488976259001,1.0652498150622016,2.060697840353612,-2.261219441515631,0.8390333282487512,0.41676809),
                     c("GO:0033690","positive regulation of osteoblast proliferation",0.0026495956790935906,5.70804304522266,-4.256118242884453,3.0318122713303706,-2.4584207560534193,0.8142485020063622,0.29250762),
                     c("GO:0034087","establishment of mitotic sister chromatid cohesion",0.011064218607861515,-5.448034436493845,2.5669373374950593,3.6522463410033232,-1.4005536242747243,0.9565911645460723,0.54197533),
                     c("GO:0034097","response to cytokine",0.3126892612355427,2.4086043994130013,5.845022133831637,5.103345247068318,-1.5991167844516372,0.9096873387874305,0.53735075),
                     c("GO:0034109","homotypic cell-cell adhesion",0.02453648835848995,-0.2052099324237584,-8.379169768877414,3.998084887936556,-2.5900668766687054,0.9961865860565648,0.00699881),
                     c("GO:0034239","regulation of macrophage fusion",0.0009637134051400874,1.7081911316142109,-5.913185302835259,2.593286067020457,-1.9726503922252434,0.8261858313962884,0.37227896),
                     c("GO:0034975","protein folding in endoplasmic reticulum",0.05576474626929534,-2.1851409943476954,6.195510923856605,4.354607782789264,-1.6149302236680652,0.9751280083555174,0.40406147),
                     c("GO:0035023","regulation of Rho protein signal transduction",0.08740215104775612,5.167716030585874,1.0217476502424316,4.549763225468786,-2.415668775632469,0.805125629930787,0.60667529),
                     c("GO:0035264","multicellular organism growth",0.00892975362358705,-4.644895227127912,-4.546070641207371,3.5591881890047756,-1.7907531512466264,0.8774562712899268,0.68804857),
                     c("GO:0035331","negative regulation of hippo signaling",0.00889278252108807,6.480285571449148,1.181084116862557,3.5573868820595074,-1.3601152580836957,0.775933151118977,0.64146493),
                     c("GO:0035556","intracellular signal transduction",4.1384737362710275,3.3286435602324116,1.6945286580058128,6.225069319444936,-1.8434508486682186,0.7615666599201651,0.47438994),
                     c("GO:0035989","tendon development",0.00047323011198694835,-2.9900044539431834,-6.083570794268916,2.285557309007774,-1.4005536242747243,0.9034284809729967,0.42634808),
                     c("GO:0035995","detection of muscle stretch",0.0015675747459567664,0.7348418283044671,5.227040713829488,2.8041394323353503,-2.744727494896694,0.9416429973224809,0.15334988),
                     c("GO:0036120","cellular response to platelet-derived growth factor stimulus",0.0037759819352291917,1.501260814702496,6.731145872830212,3.185542154854375,-1.4683933680672778,0.9366551868098749,0.62855565),
                     c("GO:0038083","peptidyl-tyrosine autophosphorylation",7.640694516455937E-05,-1.3600584751997844,7.124979182923193,1.505149978319906,-2.785156151952302,0.9797817491801223,0.12642635),
                     c("GO:0038094","Fc-gamma receptor signaling pathway",0.001355607091629279,4.246688792964552,3.5512333808535153,2.741151598851785,-2.6497519816658373,0.7934630737533875,0.42152356),
                     c("GO:0040011","locomotion",0.5148398554794674,-0.5910576194771248,-8.298003500097519,5.319903096243506,-3.161150909262745,1,-0),
                     c("GO:0040040","thermosensory behavior",0.0005890728998170868,-3.1092528114201445,-1.8819182852731517,2.380211241711606,-1.7731424297112766,0.8732571514931886,0.38755741),
                     c("GO:0042110","T cell activation",0.05417498886183919,-4.2850161666916184,-3.152646839069747,4.342047446258668,-2.612610173661271,0.8741859972147967,0.39282601),
                     c("GO:0042127","regulation of cell population proliferation",0.41876428378545044,3.2333372896507155,-0.3917538153676142,5.230201047332512,-1.6394065864347511,0.8519520107549835,0.21316029),
                     c("GO:0042157","lipoprotein metabolic process",0.29933283427274443,-2.2820016910847762,6.713891419468129,5.084386791273404,-1.453457336521869,0.9813018690027445,0.18545153),
                     c("GO:0042752","regulation of circadian rhythm",0.08089030752760239,5.912861352843011,3.6506806940094028,4.516138576717075,-2.3010299956639813,0.8732646795811407,0.14751721),
                     c("GO:0042770","signal transduction in response to DNA damage",0.1690343453655047,2.846928609766354,2.7491230693982915,4.836210145933471,-2.164943898279884,0.803113028491103,0.3170743),
                     c("GO:0043010","camera-type eye development",0.09549142827453301,-4.0258659290208545,-4.656539586136948,4.588204456038483,-1.3379981206100828,0.8322852594510054,0.68696988),
                     c("GO:0043056","forward locomotion",3.6971102498980334E-05,-1.5609770916957977,4.296270048266373,1.2041199826559248,-3.2596373105057563,0.9977713841242073,-0),
                     c("GO:0043066","negative regulation of apoptotic process",0.2887738873990356,6.048192059614752,0.44077442164083674,5.068790483415546,-3.3010299956639813,0.7710329044808444,0.57667555),
                     c("GO:0043153","entrainment of circadian clock by photoperiod",0.03754292221763123,1.9207834131045982,3.846028961121082,4.1827854420846835,-1.6354490046460282,0.8252819360703441,0.53022358),
                     c("GO:0043462","regulation of ATP-dependent activity",0.00037464050532300076,-2.2158182874727497,1.2576132179343487,2.184691430817599,-1.8010681300677909,0.8607393179692717,0.67975931),
                     c("GO:0043473","pigmentation",0.06329945695858753,-6.600344046182821,0.6380750077028623,4.409645751770205,-1.3106911408763797,1,-0),
                     c("GO:0043586","tongue development",0.0016957412346198981,-3.3037199413719573,-5.994334242701621,2.8382192219076257,-1.3745846478455919,0.8899621755814675,0.41791676),
                     c("GO:0043589","skin morphogenesis",0.0014320140367938385,-3.4637917953501347,-5.994438379687423,2.7649229846498886,-1.3745846478455919,0.8814380317762471,0.50752422),
                     c("GO:0043615","astrocyte cell migration",0.00039435842665579027,-4.316383942748791,-4.113473727551541,2.2068258760318495,-1.349498205121633,0.8720963551544011,0.46768068),
                     c("GO:0043903","regulation of biological process involved in symbiotic interaction",0.01113323133252628,5.959920545954864,5.004798540961782,3.6549462265843444,-2.596879478824182,0.886876196331815,0.12801894),
                     c("GO:0044030","regulation of DNA methylation",0.005124194806358675,0.827409227961561,-0.44940850595884924,3.3180633349627615,-2.187755303199631,0.8677549613255535,0.26698764),
                     c("GO:0044319","wound healing, spreading of cells",0.004690400537037305,-2.6485855621867938,-3.1097885171565327,3.2796669440484556,-1.3480439304669258,0.8266480561656531,0.52515336),
                     c("GO:0044403","biological process involved in symbiotic interaction",0.16708966537405834,0.16219389042288598,7.571582307767334,4.8311848528889,-1.431798275933005,0.9880404891256973,0.55838838),
                     c("GO:0044804","nucleophagy",0.11118442891526686,-2.607541631659942,2.596711736474455,4.65428245444571,-1.5151309672795976,0.9880935403595121,0.05862843),
                     c("GO:0045200","establishment of neuroblast polarity",0.0001626728509955135,-3.4968941454584526,-4.66656276589772,1.8260748027008264,-1.4005536242747243,0.8762495497154094,0.59307395),
                     c("GO:0045429","positive regulation of nitric oxide biosynthetic process",0.009225522443578894,4.471842115301869,-3.609088801410375,3.5733358400660675,-1.5816987086802545,0.8038075235723059,0.52012754),
                     c("GO:0045647","negative regulation of erythrocyte differentiation",0.0010425850904712455,6.316340359162719,-1.2521089701576293,2.6273658565927325,-2.4584207560534193,0.7381333514244722,0.60686536),
                     c("GO:0045666","positive regulation of neuron differentiation",0.019786934057454277,3.733428160765535,-4.198606646045377,3.9046614579155245,-3.070581074285707,0.7497386757215256,0.6915402),
                     c("GO:0045667","regulation of osteoblast differentiation",0.018936598699977727,1.4676756655924328,-4.332407535767551,3.885587356189656,-1.4092695942073097,0.8105317525364392,0.63408988),
                     c("GO:0045668","negative regulation of osteoblast differentiation",0.004749554301035673,6.495117075686122,-0.34704020580826933,3.285107029566812,-2.8632794328435933,0.7679313607144792,0.55965226),
                     c("GO:0045722","positive regulation of gluconeogenesis",0.001688347014120102,4.92440194874257,-4.268253389024981,2.8363241157067516,-2.563837352959244,0.8181646646137879,0.46562399),
                     c("GO:0045732","positive regulation of protein catabolic process",0.23489959683752146,4.197533824538597,-2.8712685897237495,4.97911568569024,-2.1567672219019904,0.7553273029164584,0.62289212),
                     c("GO:0045892","negative regulation of DNA-templated transcription",1.5258343712354174,5.272188998559223,-0.3589661946779675,5.791736952586863,-3.9208187539523753,0.7329483345685198,0.20416979),
                     c("GO:0046326","positive regulation of glucose import",0.004715047938703292,4.585232744999671,-5.124724263547594,3.281941933440825,-1.5431786519784014,0.7884448766033397,0.57367475),
                     c("GO:0046697","decidualization",0.0017351770772854771,-3.7810253994012974,-4.84162647475523,2.848189116991399,-1.6294869104014074,0.8571712654907278,0.41837333),
                     c("GO:0046777","protein autophosphorylation",0.001434478776960437,-1.4950149815209677,7.02638682713202,2.765668554759014,-1.9850596502070634,0.9779983934286375,0.36008171),
                     c("GO:0046827","positive regulation of protein export from nucleus",0.0019003146684475893,4.521504873671104,-5.377038578369277,2.887617300335736,-1.7999707334462298,0.7922672838197428,0.65477106),
                     c("GO:0046839","phospholipid dephosphorylation",0.14359329736587303,-3.9105414061145107,6.1939542870168856,4.765370480291648,-2.1617807780923743,0.9806833018339529,0.40330064),
                     c("GO:0048013","ephrin receptor signaling pathway",0.04839763791133186,3.1115479128611705,3.4615412670478247,4.2930751401228635,-3.0315170514460648,0.8098879189937294,0.15211288),
                     c("GO:0048102","autophagic cell death",0.0008947006804753241,-2.473954745320937,3.8960474013142226,2.561101383649056,-1.3745846478455919,0.9968136825478429,0.00573714),
                     c("GO:0048147","negative regulation of fibroblast proliferation",0.002716143663591755,7.358484800319122,1.3800783512413284,3.0425755124401905,-1.5392524581558031,0.8142315400713053,0.69690666),
                     c("GO:0048268","clathrin coat assembly",0.035341909248858605,-4.788130649694934,3.0510005547536823,4.156549151331781,-1.621783850250122,0.9590171697599038,0.25392788),
                     c("GO:0048522","positive regulation of cellular process",4.210320912127379,4.481546705312644,-2.5631008129624973,6.232544310796654,-1.834755673874689,0.7389049029085789,0.69176276),
                     c("GO:0048525","negative regulation of viral process",0.018221824051664108,7.163626219665157,1.5949138945374148,3.868879446237088,-1.7785857621576613,0.8189301800670925,0.44586228),
                     c("GO:0048661","positive regulation of smooth muscle cell proliferation",0.005156236428524457,5.600336238981788,-4.027124838367988,3.3207692283386865,-2.1163385648463824,0.8065283149156078,0.59012891),
                     c("GO:0048920","posterior lateral line neuromast primordium migration",0.0004658358914871522,-4.122625240845286,-4.178498972008364,2.278753600952829,-1.917933065714887,0.8656870326210248,0.47114954),
                     c("GO:0050680","negative regulation of epithelial cell proliferation",0.018283442555829078,6.876619955215142,0.9315206528921198,3.8703453710809597,-2.2055119533408303,0.7952680255785839,0.56960666),
                     c("GO:0050688","regulation of defense response to virus",0.013716279027121706,6.273632061164739,2.734257760935795,3.745543201998024,-1.493765640387874,0.8320237231041857,0.62019347),
                     c("GO:0050764","regulation of phagocytosis",0.014965902291587242,3.1786103652944266,-5.907705196315898,3.7834032811225633,-1.4590451910738673,0.8045237487819984,0.60788752),
                     c("GO:0050821","protein stabilization",0.10143638155636905,5.680130189843421,4.088150417141459,4.614433158550452,-2.515700160653214,0.8657735103293462,0.34813705),
                     c("GO:0050853","B cell receptor signaling pathway",0.020541144548433473,4.095392164014354,0.05462385494172586,3.920905604164024,-2.417936637088291,0.6992666888967743,0.69477021),
                     c("GO:0050872","white fat cell differentiation",0.0015996163681225493,-3.1787076584206857,-6.3123767125999,2.8129133566428557,-1.9500071430798573,0.9053960587321584,0.25349818),
                     c("GO:0051054","positive regulation of DNA metabolic process",0.08397123273585075,4.286933737779803,-3.0201145891676324,4.532372133567877,-1.6929320493387017,0.7764883402154747,0.68856099),
                     c("GO:0051091","positive regulation of DNA-binding transcription factor activity",0.0014049018949612527,-0.15198372956199765,0.44153989006191074,2.756636108245848,-3.638272163982407,0.8290056256669621,0.25521058),
                     c("GO:0051153","regulation of striated muscle cell differentiation",0.01946651783579645,1.473128714538747,-4.218191993727335,3.8975721138257304,-2.424812155072339,0.8065523271861994,0.60355573),
                     c("GO:0051240","positive regulation of multicellular organismal process",0.4722097095579764,4.862411560354065,-3.034911525936926,5.282366036998229,-1.3390397082239165,0.7123048637374481,0.68769205),
                     c("GO:0051301","cell division",1.5693197819947182,-3.764573481067649,0.9624085241236819,5.803940988740412,-1.7464197104378172,0.9949229042364895,0.01106416),
                     c("GO:0051382","kinetochore assembly",0.029438856549854744,-5.1329917378080046,2.937452953541807,4.077186154085897,-1.4005536242747243,0.9458756687630265,0.66203517),
                     c("GO:0051451","myoblast migration",0.0037883056360621856,-6.8678960349007925,-1.9786105723952725,3.1869563354654122,-2.6458915608525992,0.9865229260233396,0.00622631),
                     c("GO:0051592","response to calcium ion",0.07100916419970824,2.6064319881283904,6.043819076011174,4.459558332454132,-2.404503778174426,0.9229283207831662,0.42163898),
                     c("GO:0051726","regulation of cell cycle",0.9132256675674799,3.7358432895319647,-0.6909527191643022,5.568808139000413,-1.625434939277235,0.8436753670898535,0.23187904),
                     c("GO:0051894","positive regulation of focal adhesion assembly",0.003687251289231639,4.339982942549964,-4.408758627705486,3.1752218003430523,-3.853871964321762,0.7928437199747166,0.12780185),
                     c("GO:0051897","positive regulation of phosphatidylinositol 3-kinase/protein kinase B signal transduction",0.026927286320090678,5.254608438913907,-2.602953673807746,4.038461196178564,-1.7991495019089225,0.7603601043299384,0.65449542),
                     c("GO:0051946","regulation of glutamate uptake involved in transmission of nerve impulse",0.00019224973299469776,2.1761968413060253,-7.917672114717849,1.8976270912904414,-2.0958256317158366,0.8638172368650963,0.3927646),
                     c("GO:0060039","pericardium development",0.002023551676777524,-3.948399626144404,-4.91771844943768,2.9148718175400505,-1.4564285760376345,0.8415018487125195,0.58234935),
                     c("GO:0060048","cardiac muscle contraction",0.03486128491637186,-5.006094190684462,-3.6219446416062584,4.15060295179301,-2.838631997765025,0.8864386049168095,-0),
                     c("GO:0060049","regulation of protein glycosylation",0.0006654798449816461,-0.07964522848423816,-0.2802326133186556,2.432969290874406,-1.349498205121633,0.8645511262056738,0.2123578),
                     c("GO:0060083","smooth muscle contraction involved in micturition",0.00021196765432748727,-5.2772609660053975,-3.468566203102432,1.9395192526186185,-1.7731424297112766,0.9113751451766726,0.63419309),
                     c("GO:0060155","platelet dense granule organization",0.002220730890105419,-5.675738780664557,3.418049787014735,2.9552065375419416,-1.9685915357483759,0.9653420275682121,0.29134061),
                     c("GO:0060297","regulation of sarcomere organization",0.0010968093741364165,1.8245436072189642,-7.059731849056302,2.649334858712142,-2.0963674839157624,0.863052283627425,0.43997441),
                     c("GO:0060318","definitive erythrocyte differentiation",0.00017499655182850692,-3.7635624123629903,-3.623783042492588,1.8573324964312685,-1.9726503922252434,0.8690578983595567,0.44474949),
                     c("GO:0060337","type I interferon-mediated signaling pathway",0.0090234137499178,2.4012174281120187,4.387456823982517,3.5637183399656776,-2.416801226031377,0.7771752577113599,0.54891864),
                     c("GO:0060379","cardiac muscle cell myoblast differentiation",0.0010647677519706336,-4.0883178983166495,-5.101217742980099,2.6364878963533656,-2.416801226031377,0.8369421456834444,0.56498298),
                     c("GO:0060419","heart growth",0.004185128802884574,-4.1683824043121485,-4.943923218277859,3.230193378869046,-3.3279021420642825,0.8385469383926257,0.38130142),
                     c("GO:0060947","cardiac vascular smooth muscle cell differentiation",0.0005249896554855208,-4.060822364813073,-5.18583233693628,2.330413773349191,-1.9726503922252434,0.8383192175898901,0.6955647),
                     c("GO:0060996","dendritic spine development",0.007162534924135791,-4.416141171845281,-3.041059810066495,3.4634450317704277,-1.5529971015338375,0.8349623390878654,0.40829561),
                     c("GO:0061026","cardiac muscle tissue regeneration",0.0005348486161519156,-3.008871485068765,-6.3156657036654495,2.3384564936046046,-1.6899442622491085,0.9089041836473462,0.60087667),
                     c("GO:0061036","positive regulation of cartilage development",0.007643159256622535,4.009956332973457,-4.305403600009558,3.4916417934775863,-1.6176226965318863,0.7367840695655133,0.59036775),
                     c("GO:0061056","sclerotome development",0.00043132952915477056,-3.7319656579884524,-5.083781551687657,2.24551266781415,-1.7731424297112766,0.860383141586791,0.54974695),
                     c("GO:0061952","midbody abscission",0.004288647889881719,-5.223439000378007,2.117465982183231,3.240798771117331,-1.6149302236680652,0.9619000638183391,0.5133694),
                     c("GO:0065008","regulation of biological quality",3.3795629857941245,3.7620407233916064,-0.32423390984608846,6.137089719032985,-2.4461169733561254,0.8436214361847177,0.24032368),
                     c("GO:0070345","negative regulation of fat cell proliferation",0.00032288096182442826,7.57728543151052,1.8557854660986275,2.12057393120585,-2.031984286006358,0.8313997853265118,0.62706343),
                     c("GO:0070372","regulation of ERK1 and ERK2 cascade",0.09023660223934461,5.206359157658406,1.076993434359664,4.563623453690032,-1.4563043676907552,0.8058398306209158,0.64649001),
                     c("GO:0070507","regulation of microtubule cytoskeleton organization",0.12544295077904027,2.720609595217512,-1.8168919212980272,4.706683651763288,-2.7351821769904636,0.8237598023174638,0.59028651),
                     c("GO:0070571","negative regulation of neuron projection regeneration",0.00308338994841496,5.682718483521063,-0.7933584998367531,3.097604328874411,-1.8683813356508745,0.7241150158571654,0.66846157),
                     c("GO:0070849","response to epidermal growth factor",0.00823469689660622,1.5040480090960777,6.59850848166777,3.5240064455573727,-2.243363891754152,0.9348574035567185,0.49391738),
                     c("GO:0070885","negative regulation of calcineurin-NFAT signaling cascade",0.0032805691617428553,6.640452011468584,1.3839436401504779,3.1245042248342823,-2.2549252084179425,0.7867201344831952,0.60789009),
                     c("GO:0071103","DNA conformation change",0.8233982121957907,-5.409486807316305,2.866366227293501,5.523840077089552,-2.6819366650372385,0.9491131945341079,0.64824799),
                     c("GO:0071285","cellular response to lithium ion",0.001252088004632134,3.164850305378565,6.161363719169918,2.7067177823367587,-1.3019864960608183,0.9280613026586062,0.61470049),
                     c("GO:0071287","cellular response to manganese ion",0.0003425988831572178,3.2937203778204487,6.265673569244706,2.146128035678238,-1.4005536242747243,0.9322106277265987,0.57453476),
                     c("GO:0071300","cellular response to retinoic acid",0.00979487742206319,2.74821360906296,6.0731297986287744,3.5993371329924893,-1.889410289700751,0.9091549865195021,0.67807418),
                     c("GO:0071318","cellular response to ATP",0.0011510336578015878,2.453930873888851,6.385276976448439,2.670245853074124,-1.8467950999157157,0.9113109309471071,0.62821184),
                     c("GO:0071346","cellular response to type II interferon",0.007618511854956549,1.7942194521771502,5.43666913277233,3.4902394852462875,-1.5395532161192795,0.8892421998362429,0.63225951),
                     c("GO:0071347","cellular response to interleukin-1",0.011308227884354786,2.8045971483372125,5.9441798113089,3.6617180576946593,-1.7070797003999938,0.9136069531714784,0.64835115),
                     c("GO:0071549","cellular response to dexamethasone stimulus",0.013531423514626804,2.3559293812916473,6.147518959714497,3.7396514437093766,-2.632644078973981,0.8968388589668385,0.20468023),
                     c("GO:0071806","protein transmembrane transport",0.7053568761370463,-7.140373968585767,-0.5883162616157072,5.456639279342896,-1.819874124835946,0.9702301515576653,0.5469966),
                     c("GO:0072102","glomerulus morphogenesis",0.0007862521131449819,-3.6712235646597358,-4.984055145798063,2.505149978319906,-1.3019864960608183,0.8654856728569672,0.40320584),
                     c("GO:0072659","protein localization to plasma membrane",0.1961243045365909,-7.046687754707955,-0.44103539436227107,4.9007657318060245,-2.452225294612177,0.9725075056332079,0.00827286),
                     c("GO:0086047","membrane depolarization during Purkinje myocyte cell action potential",0.00023661505599347418,4.872749618291771,3.3318642468251647,1.9867717342662448,-2.326058001365912,0.8181663614110702,0.49330786),
                     c("GO:0090069","regulation of ribosome biogenesis",0.026670953342764415,1.8695532107386619,0.6211789198289667,4.034307529596563,-1.4285240963180563,0.8675523335606792,0.54629382),
                     c("GO:0090084","negative regulation of inclusion body assembly",0.0008232232156439621,7.313269780695088,0.24329184236994977,2.525044807036845,-1.4005536242747243,0.8133636898056985,0.45959416),
                     c("GO:0090087","regulation of peptide transport",0.054261254767670145,3.2318860406335235,-6.4971499038256955,4.3427384165554175,-2.0609802235513337,0.8288133657528854,0.5149342),
                     c("GO:0090091","positive regulation of extracellular matrix disassembly",0.001547856824623977,4.348809826474831,-4.824175780275333,2.798650645445269,-1.8664610916297826,0.8098411200895973,0.5063792),
                     c("GO:0090108","positive regulation of high-density lipoprotein particle assembly",0.00018485551249490168,4.614417686991208,-4.369596667068652,1.8808135922807914,-1.3745846478455919,0.7764211839031387,0.53225131),
                     c("GO:0090210","regulation of establishment of blood-brain barrier",0.00023907979616007285,0.8000278136284228,-6.934952462456192,1.9912260756924949,-2.0958256317158366,0.8451007220930344,0.52587948),
                     c("GO:0090263","positive regulation of canonical Wnt signaling pathway",0.10080294333355319,5.055383620328485,-2.415594237984201,4.611712689430539,-2.348721986001856,0.7445509507784617,0.61343594),
                     c("GO:0090336","positive regulation of brown fat cell differentiation",0.004638640993538733,3.5943922377185795,-4.681713280391931,3.274850320016665,-1.7110803943382735,0.7682147468763451,0.63987477),
                     c("GO:0090650","cellular response to oxygen-glucose deprivation",3.4506362332381646E-05,1.9815050361835584,5.674304023255144,1.1760912590556813,-1.9726503922252434,0.9272574968925307,0.5035789),
                     c("GO:0097094","craniofacial suture morphogenesis",0.0015084209819583977,-3.881936916148712,-5.10649588712042,2.787460474518415,-1.5151309672795976,0.8560176730571677,0.41560739),
                     c("GO:0097164","ammonium ion metabolic process",0.08456030563566783,-5.082921254198824,5.339146602637265,4.535408059915369,-1.8127613801685214,0.9917703471364988,0.07513778),
                     c("GO:0097167","circadian regulation of translation",0.00037464050532300076,-0.27076673200363505,-0.08799037533723922,2.184691430817599,-1.8184422261372137,0.8719903718744834,0.69325215),
                     c("GO:0097753","membrane bending",0.00025633297732626366,-5.243084537447984,3.9012164876329964,2.0211892990699383,-1.917933065714887,0.9725993521105081,0.18937677),
                     c("GO:0098719","sodium ion import across plasma membrane",0.0490631177563135,-7.2541957392846,-0.9463649134216807,4.299005816450637,-1.6899442622491085,0.9763195604494108,0.47725008),
                     c("GO:0098771","inorganic ion homeostasis",1.0570555799893464,-4.863511125829961,-0.08702363001917032,5.6323276931440835,-1.5395532161192795,0.99038571272706,0.49482653),
                     c("GO:0099084","postsynaptic specialization organization",0.00580939257267311,-5.851585118875246,2.5789685057941196,3.3725438007590705,-1.9783972839717578,0.9581899632158491,0.30857275),
                     c("GO:0120162","positive regulation of cold-induced thermogenesis",0.007579076012290969,4.905274705832891,-3.6015740221034003,3.4879863311293935,-2.485452247339714,0.7557761491230395,0.4691014),
                     c("GO:0150007","clathrin-dependent synaptic vesicle endocytosis",0.00471997741903649,-7.07120656522231,-0.8870658186132223,3.2823955047425257,-2.031984286006358,0.9692506054140846,0.32908148),
                     c("GO:0150093","amyloid-beta clearance by transcytosis",0.0007024509474806264,-5.434952393882046,-3.0390758993826292,2.456366033129043,-1.5726762136427528,0.8998990870846121,0.30673485),
                     c("GO:1900221","regulation of amyloid-beta clearance",0.0026298777577608012,1.455768070902934,-5.267888857973544,3.0285712526925375,-1.5055671012736014,0.8287555898462837,0.3845541),
                     c("GO:1900246","positive regulation of RIG-I signaling pathway",0.0011855400201339695,5.798151834317753,-3.0685656187542625,2.6830470382388496,-2.1655792963184672,0.7792444142761313,0.49749183),
                     c("GO:1901223","negative regulation of non-canonical NF-kappaB signal transduction",0.01620566659538638,6.375319381470008,1.0146117450765926,3.817961804531994,-2.7166987712964503,0.7688960912141426,0.53678803),
                     c("GO:1901897","regulation of relaxation of cardiac muscle",0.00021196765432748727,2.730707963631502,-7.354014889804261,1.9395192526186185,-2.769551078621726,0.8409222499181823,0.1012796),
                     c("GO:1901990","regulation of mitotic cell cycle phase transition",0.2846208002183168,3.0915682864713454,0.1590970264199983,5.062499253590347,-1.9775716288145135,0.8488861703427665,0.20496548),
                     c("GO:1902065","response to L-glutamate",0.00043132952915477056,2.7935914171257177,6.588023160683968,2.24551266781415,-1.5802094138936373,0.9320285303141898,0.41687571),
                     c("GO:1902218","regulation of intrinsic apoptotic signaling pathway in response to osmotic stress",0.0006704093253148435,6.6852920279831025,3.0961626340029325,2.436162647040756,-2.0958256317158366,0.8358246821624485,0.44080619),
                     c("GO:1902624","positive regulation of neutrophil migration",0.005419963626350517,5.451847140114295,-4.221964092396207,3.342422680822206,-1.5808706922580242,0.7964521054048458,0.65485146),
                     c("GO:1902685","positive regulation of receptor localization to synapse",0.0002489387568264676,4.3347294742208184,-5.983677719435825,2.0086001717619175,-2.326058001365912,0.8287202002638083,0.39707888),
                     c("GO:1902746","regulation of lens fiber cell differentiation",0.0007517457508126002,1.7955142027722824,-5.430758239210488,2.48572142648158,-1.6149302236680652,0.7934811851026367,0.51107125),
                     c("GO:1902894","negative regulation of miRNA transcription",0.0023020673156031756,7.049248039711734,0.7718072200231861,2.9708116108725178,-1.7110803943382735,0.8163052236755886,0.55401842),
                     c("GO:1903077","negative regulation of protein localization to plasma membrane",0.0024376280247661035,7.398401307685791,-0.39876512070112513,2.99563519459755,-2.4282911681913126,0.7914387189042571,0.39069153),
                     c("GO:1903146","regulation of autophagy of mitochondrion",0.007231547648800554,1.9252682051469328,-2.1694508270083253,3.467608105583633,-2.2636034977233574,0.8335428563624755,0.41028305),
                     c("GO:1903393","positive regulation of adherens junction organization",0.001663699612454115,4.457374991732439,-4.782024750275506,2.829946695941636,-3.4436974992327127,0.8091939393741864,0.50810213),
                     c("GO:1903428","positive regulation of reactive oxygen species biosynthetic process",0.0012816648866313183,5.0227149803113145,-4.471366637799971,2.7168377232995247,-1.5550188879120552,0.8227112619966823,0.45738836),
                     c("GO:1903955","positive regulation of protein targeting to mitochondrion",0.0016144048091221415,4.589457490635611,-5.414094500981602,2.8169038393756605,-1.8975662943186637,0.8004446188196764,0.55503423),
                     c("GO:1904018","positive regulation of vasculature development",0.03908338482175541,4.156676365750278,-3.909936024369004,4.200248413515354,-1.5775743236287953,0.712987705354953,0.6461048),
                     c("GO:1904181","positive regulation of membrane depolarization",0.001949609471779563,5.402133661564282,-4.683588254752507,2.8987251815894934,-1.5162701009999762,0.8254156382105582,0.36901543),
                     c("GO:1904526","regulation of microtubule binding",4.190058283217772E-05,-2.4538139093979057,1.2317392051246694,1.255272505103306,-3.00436480540245,0.8699815058300687,0.62217145),
                     c("GO:1904591","positive regulation of protein import",0.00017499655182850692,4.515114244293133,-5.814316284436679,1.8573324964312685,-1.5991167844516372,0.814623400480012,0.59047947),
                     c("GO:1904754","positive regulation of vascular associated smooth muscle cell migration",0.0008306174361437582,5.418237537460254,-5.155277916137576,2.5289167002776547,-1.8632794328435933,0.831321239147431,0.2770188),
                     c("GO:1905150","regulation of voltage-gated sodium channel activity",5.9153763998368544E-05,1.6789962866697616,-7.81583283070351,1.3979400086720377,-1.5151309672795976,0.8240368116365862,0.63058072),
                     c("GO:1990776","response to angiotensin",0.007130493301970008,2.238773286850272,6.36820905907497,3.4614985267830187,-1.7344746647809262,0.9125391391518636,0.53259764),
                     c("GO:1990791","dorsal root ganglion development",0.000611255561316475,-3.794651367580431,-4.5072338875963105,2.3961993470957363,-1.484788695672198,0.871709873716915,0.43109911),
                     c("GO:1990830","cellular response to leukemia inhibitory factor",0.013802544932952658,2.711158654104535,5.91295465647505,3.748265572668741,-1.9389246763702082,0.9126864644554976,0.64031112),
                     c("GO:1990845","adaptive thermogenesis",0.005651649202010794,-3.2117111201636512,4.419205016548518,3.360593413565249,-1.495529137505581,0.996578703162776,0.02717669),
                     c("GO:2000739","regulation of mesenchymal stem cell differentiation",0.000998219767472469,0.9267672420105437,-6.3519295104613995,2.6085260335771943,-1.8664610916297826,0.8344228060488398,0.5179882),
                     c("GO:2000809","positive regulation of synaptic vesicle clustering",0.00032288096182442826,4.432682726226004,-5.903634765346999,2.12057393120585,-1.325322532126801,0.8200349623976002,0.40151772),
                     c("GO:2000973","regulation of pro-B cell differentiation",0.0007665341918121924,0.9000716433310285,-6.506120819712818,2.494154594018443,-1.7731424297112766,0.8379859658558476,0.51154051));

## Parse data ##
revigo.dat.parse <- function(revigo.data, revigo.names){
  one.data <- data.frame(revigo.data)
  names(one.data) <- revigo.names
  one.data <- one.data [(one.data$plot_X != "null" & one.data$plot_Y != "null"), ]
  one.data$plot_X <- as.numeric( as.character(one.data$plot_X) )
  one.data$plot_Y <- as.numeric( as.character(one.data$plot_Y) )
  one.data$log_size <- as.numeric( as.character(one.data$log_size) )
  one.data$value <- as.numeric( as.character(one.data$value) )
  one.data$frequency <- as.numeric( as.character(one.data$frequency) )
  one.data$uniqueness <- as.numeric( as.character(one.data$uniqueness) )
  one.data$dispensability <- as.numeric( as.character(one.data$dispensability) )
  return(one.data)
}

BN.mfuzz.cluster6.one.data <- revigo.dat.parse(BN.mfuzz.cluster6.revigo.data, BN.mfuzz.cluster6.revigo.names)
BN.mfuzz.cluster5.one.data <- revigo.dat.parse(BN.mfuzz.cluster5.revigo.data, BN.mfuzz.cluster5.revigo.names)

BA.mfuzz.cluster5.one.data <- revigo.dat.parse(BA.mfuzz.cluster5.revigo.data, BA.mfuzz.cluster5.revigo.names)

BA.snp.one.data <- revigo.dat.parse(BA.snp.revigo.data, BA.snp.revigo.names)
BN.snp.one.data <- revigo.dat.parse(BN.snp.revigo.data, BN.snp.revigo.names)

#convert p value and size columns 
BN.mfuzz.cluster6.one.data$value  <- as.numeric(BN.mfuzz.cluster6.one.data$value %>% lapply(function(x) 10^x))
BN.mfuzz.cluster5.one.data$value  <- as.numeric(BN.mfuzz.cluster5.one.data$value %>% lapply(function(x) 10^x))

BA.mfuzz.cluster5.one.data$value  <- as.numeric(BA.mfuzz.cluster5.one.data$value %>% lapply(function(x) 10^x))

BA.snp.one.data$value  <- as.numeric(BA.snp.one.data$value %>% lapply(function(x) 10^x))
BN.snp.one.data$value  <- as.numeric(BN.snp.one.data$value %>% lapply(function(x) 10^x))

## Plot ##  
plot.revigo <- function(one.data){
  ex <- one.data [ one.data$dispensability < 0.35, ]
  one.x_range = max(one.data$plot_X) - min(one.data$plot_X)
  one.y_range = max(one.data$plot_Y) - min(one.data$plot_Y)
  p1 <- ggplot( data = one.data ) + geom_point( aes( plot_X, plot_Y, color = value, size = log_size), alpha = I(0.8) ) + 
    scale_colour_viridis(option = "viridis", limits = c(0, 0.05), name = "q-value" , direction = -1) + 
    geom_point( aes(plot_X, plot_Y, size = log_size), shape = 21, fill = "transparent", colour = I (alpha ("black", 0.6) )) + 
    theme_bw() + geom_text_repel(data = ex, aes(plot_X, plot_Y, label = description), colour = I(alpha("black", 0.85)), size = 4, 
                                  min.segment.length = Inf, max.overlaps = 20) +
    labs (y = "Semantic space y", x = "Semantic space x") + theme(legend.key = element_blank()) + 
    xlim(min(one.data$plot_X)-one.x_range/10,max(one.data$plot_X)+one.x_range/10) + 
    ylim(min(one.data$plot_Y)-one.y_range/10,max(one.data$plot_Y)+one.y_range/10) + 
    scale_size_continuous(breaks = c(1, 3, 5, 7), name = "Log Size", limits = c(0, 7)) + 
    theme(axis.title = element_text(size = 14), axis.text = element_text(size = 11), legend.text = element_text(size=11), 
          legend.position = "right")
  return(p1)
}

BN.mfuzz.cluster6.plot <- plot.revigo(BN.mfuzz.cluster6.one.data)
BN.mfuzz.cluster5.plot <- plot.revigo(BN.mfuzz.cluster5.one.data)

BA.mfuzz.cluster5.plot <- plot.revigo(BA.mfuzz.cluster5.one.data)

BA.snp.plot <- plot.revigo(BA.snp.one.data)
BN.snp.plot <- plot.revigo(BN.snp.one.data)

#save plots 
ggsave("BN_StFvsSwF_up_REVIGO.tiff", plot = BN.StFvsSwF.up.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_StFvsSwF_down_REVIGO.tiff", plot = BN.StFvsSwF.down.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_SwFvsSwP_down_REVIGO.tiff", plot = BN.SwFvsSwP.down.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_vennGenes_REVIGO.tiff", plot = BN.venngenes.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_mfuzz_cluster6_0.35_REVIGO.tiff", plot = BN.mfuzz.cluster6.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_mfuzz_cluster5_0.35_REVIGO.tiff", plot = BN.mfuzz.cluster5.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_30perc_REVIGO.tiff", plot = BN.30perc.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BN_sameDir_up_0.35_REVIGO.tiff", plot = BN.sameDir.up.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)

ggsave("BA_StFvsSwF_up_REVIGO.tiff", plot = BA.StFvsSwF.up.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BA_StFvsSwF_down_REVIGO.tiff", plot = BA.StFvsSwF.down.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BA_SwFvsSwP_up_REVIGO.tiff", plot = BA.SwFvsSwP.up.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BA_SwFvsSwP_down_REVIGO.tiff", plot = BA.SwFvsSwP.down.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BA_mfuzz_cluster5_0.35_REVIGO.tiff", plot = BA.mfuzz.cluster5.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)
ggsave("BA_30perc_REVIGO.tiff", plot = BA.30perc.plot, device = "tiff", width = 10, height = 8, units = "in", dpi = 600)

ggsave("BA_snp_REVIGO.tiff", plot = BA.snp.plot, device = "tiff", width = 10, height = 10, units = "in", dpi = 600)
ggsave("BN_snp_REVIGO.tiff", plot = BN.snp.plot, device = "tiff", width = 10, height = 10, units = "in", dpi = 600)

# Panel BN plots 
panel.plot <- (BN.mfuzz.cluster5.plot +  BN.mfuzz.cluster6.plot) / BA.mfuzz.cluster5.plot +
                  plot_layout(guides = "collect") + plot_annotation(tag_levels = "A")
panel.plot

ggsave("REVIGO_panel.tiff", plot = panel.plot, device = "tiff", width = 11, height = 7, units = "in", dpi = 600)
